Allow corner radius to be set dynamically
